requirements.txt
# ========== BBS图片爬虫核心依赖 ==========

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0  # 异步HTTP支持
httpx>=0.25.0  # 现代HTTP客户端

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0
parsel>=1.8.0  # 更强大的选择器

# 动态页面处理（可选，按需启用）
# playwright>=1.40.0
# selenium>=4.15.0

# 图片处理
Pillow>=10.0.0
imagehash>=4.3.1  # 图片去重

# 反爬虫
fake-useragent>=1.4.0
pycryptodome>=3.19.0  # 加密支持

# 数据存储
pymongo>=4.6.0  # MongoDB
redis>=5.0.0

# 任务调度
apscheduler>=3.10.0

# 工具库
loguru>=0.7.0  # 日志
tqdm>=4.66.0  # 进度条
tenacity>=8.2.0  # 重试机制
python-dotenv>=1.0.0  # 环境变量

# 数据验证
pydantic>=2.5.0

# 文件操作
aiofiles>=23.2.0  # 异步文件IO
